% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_radar_stats}
\alias{create_radar_stats}
\title{Create radar chart for selected pokemon}
\usage{
create_radar_stats(pokemon)
}
\arguments{
\item{pokemon}{The selected pokemon data.}
}
\value{
A chart htmlwidget.
}
\description{
Create radar chart for selected pokemon
}
