% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_max_of_max}
\alias{get_max_of_max}
\title{Get max of all stats maxes}
\usage{
get_max_of_max(data = get_max_stats())
}
\arguments{
\item{data}{Returned by \link{get_max_stats} with stat
set to NULL.}
}
\value{
A numeric value.
}
\description{
Needed by \link{create_radar_stats} to get the max
of the radar chart axis.
}
\keyword{internal.}
