% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_max_stats}
\alias{get_max_stats}
\title{Get max of each stat}
\usage{
get_max_stats(stat = NULL)
}
\arguments{
\item{stat}{Optional stat column to select.}
}
\value{
If stat is NULL, a 1 row tibble is returned containing
the max of each stat among all pokemons. If a stat is specified,
a numeric value is returned instead.
}
\description{
Loop over poke_data and get the max of each stat.
}
\keyword{internal}
