% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_type_colors}
\alias{get_type_colors}
\title{Set up colors according to the pokemon type}
\usage{
get_type_colors(type)
}
\arguments{
\item{type}{Pokemon type.}
}
\value{
A color, string.
}
\description{
Set up colors according to the pokemon type
}
\keyword{internal}
