% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{other_stats_names}
\alias{other_stats_names}
\title{Stats names are all the same for any pokemon ...}
\usage{
other_stats_names()
}
\value{
A vector of stat names.
}
\description{
Stats names are all the same for any pokemon ...
}
\keyword{internal}
