% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_pokemon_stats}
\alias{process_pokemon_stats}
\title{Internal function needed by \link{create_radar_stats}.}
\usage{
process_pokemon_stats(stats)
}
\arguments{
\item{stats}{List of stats from selected pokemon.}
}
\value{
A dataframe.
}
\description{
Rearrange stats for echart4r...
}
\keyword{internal}
