% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_pokemon}
\alias{select_pokemon}
\title{Function to select pokemon}
\usage{
select_pokemon(selected)
}
\arguments{
\item{selected}{Selected pokemon name.}
}
\value{
A list containing the selected pokemon data.
}
\description{
Select pokemon in the \link{poke_data}.
}
\keyword{internal}
